refactor(website): client-side render spline components
